+++ CHANGELOG.md
@@ -4,7 +4,9 @@ FaasJS use [Semantic Versioning](https://semver.org/).
 
 ## Stable
 
-[`v5.0.0 (2024-12-29)`](https://github.com/faasjs/faasjs/compare/v4.7.2...v5.0.0-beta.1)
+[`v5.0.0 (2025-01-02)`](https://github.com/faasjs/faasjs/compare/v4.7.2...v5.0.0)
+
+Happy New Year! 🎉
 
 - [Break] Convert all packages to esm.
 - [Break] Remove `@faasjs/mongo`, `@faasjs/vue-plugin`.
@@ -16,7 +18,7 @@ FaasJS use [Semantic Versioning](https://semver.org/).
 
 - `@faasjs/logger`
   - [Break] Move `@faasjs/logger/transport` to `@faasjs/logger`.
-  - [Feature] Add reset method to logger transport.
+  - [Break] Move all transports api to `getTransport`.
 
 [`v4.7.2 (2024-12-25)`](https://github.com/faasjs/faasjs/compare/v4.7.0...v4.7.2)